feat: implement function to update note, and refactor some code
Firstly, change the class of first span in note-options. Add indentation
to the two spans for better readability. Add a click evenlistener to
the edit-note span and have createNote as a cb function with the
notes id as an argument in the parameter. Also add a function
named updateNote which takes the note id as an argument. Create an object
within, with appropriate properties to update a row in "notes" table.
And then add a PATCH request and sent the noteData.
And change createNote to accept id as argument in the parameter.
Also add an if statement that checks if id is passed through as an arg,
if not, then make the saveNoteButton have addNoteToDatabase as a click eventlistener
cb function. But if there is an id, fetch the specific note and fill in the
title and content in the editor so the user can see their current content.
And add updateNote as a cb function to saveNoteButton to update note in the db.