fix: remove last child if only 2 exist, and fix cb function
Prior to this commit, the last child would get removed with each
click of the button, which can make the mainContentDiv empty. But the top bar
should always be present. Also, the createNote callback function didn't
specify an arg in the parameter, which would automatically add
the pointer event in the argument. The fix is to add the cb function
inside another cb function which in this case is an arrow function.
This makes it so we can specify that we are not adding an arg and thus
avoiding the event object passing through the parameter.